[BREAK][FEATURE] Adds image variants selector to content element
So far pizpalue had an additional image variants definition that was
automatically selected when no frame has been used in the content
element. This works fine as long as the no frame is used for content
elements that weren't nested. To be more flexible the variants selector
has been introduced.

**Corrective action:**

Since the variant isn't assigned automatically any more image content
elements spanning the page width might be rendered with a low
resolution hence look not sharp. To correct it these content elements
need to be revised manually.