[TASK] Mark content element background image deprecated
The bootstrap package offers a background image field for
content elements. This makes the field `tx_pizpalue-bgmedia`
obsolete.

The bootstrap package configuration has been adapted to enable
the background image as well when no frame is used for the
content element. In this case just a wrapping div is added.