[FEATURE] Add class `pp-extend-link`
Allows to extend a link to an ancestor element.
Assign the class `pp-extend-link` to an a-tag
to link the same url to the closest ancestor
having the class `frame-container` assigned to.
The target ancestor container might be overwritten
by assigning the class `ppc-el-[ancestor class name]`
to the a-tag. As an example when assigning the classes
`pp-extend-link ppc-el-up-my-container` to an a-tag
the closest ancestor from the link having the class
`up-my-container` assigned to will be linked to
the same url as defined in the a-tag.