[!!!][BUGFIX] Correct content element css classes
Description
-----------

The bootstrap package introduced the field `frame_layout`.
To distinguish classes related to the `layout` field the
prefix has been changed from `frame-layout` to `layout-`.
The prefix `layout-` is not added to tile layouts.

ATTENTION: Classes related to the field `frame_layout`
are now prefixed with `frame-layout`.

Corrective action
-----------------

Adapt the style definitions in the site package if the
class `frame-layout-pp-emphasize-media` has been used.