[FEATURE] Add scroll menu using bootstrap scrollspy
Content elements have been equipped with fields allowing
them to be used as menu items in a scroll menu. The scroll
menu itself can form part of the main menu, substitute
the main menu or be integrated into a template. With the
easyconf form one pager websites can be defined by just
selecting a main page.

ATTENTION: The database structure needs to be updated.