[!!!][FEATURE] Add scss frame background variables
Description
-----------

To simplify frame background adjustments scss
variables `pp-frame-background-[color-key]`
have been added where `[color-key]` has to be
substituted with one of the brand colors
(primary, secondary, etc.).

To adjust a frame with a background color
a typoscript declaration can be added to
the site package as following:

.. code-block:: typoscript

   plugin.bootstrap_package.settings.scss {
      pp-frame-background-primary = ('color': 'red')
   }

With this change the link styling has been
simplified by dropping the extended styling
when `link-color: inherit` has been in place.
In case a site package makes use of it the
link styling will be altered.

Corrective actions
------------------

Redefine the link styling in the site package.