[FEATURE] Add scss panel classes
Description
-----------

The classes define background and text colors
from a container by using the settings defined
for frames. Elements can be nested while
maintaining consistent background and text
colors. Use the classes as following:

.. code-block:: css

   pp-panel pp-panel-primary

This would create a panel with the same
background and text colors as they are used
for the frame with primary background color.
You might replace the word `primary` in the
above example by any other brand color.