[FEATURE] Trick link handler email verification
Description
~~~~~~~~~~~

The email link dialog validates the email address
before returning to the editor. To be able to use
constants as place holders for email addresses some
at-text (e.g. @a) can be appended to the constant
(e.g. ###myEmail###@a) allowing the link dialog
to be closed. Upon saving the content the at-text
will get removed.