[FEATURE] Add conditional image reload url segment
Description
~~~~~~~~~~~

The field `Reload url segment` has been added to
image content elements. When activated a cache
dependent url segment is added to the image url.
This ensures browsers reload the image even if its
name didn't change on the server. A typical use
case is integrating webcam images where the latest
snapshot is pulled in with FTP and rendered with
an image content element. HINT: When pulling in
the latest image the processed images need to be
deleted too.